modularize schedule parsing functions
