+++ client/src/components/Schedule.jsx
@@ -2,37 +2,10 @@ import { useState, useEffect } from 'react';
 import '../styles/Schedule.css';
 import scheduleData from '../data/schedule.json';
 import routeData from '../data/routes.json';
+import { parseSchedule } from '../data/parseSchedule';
 
 export default function Schedule() {
-    const weeklySchedule = [
-        'SUNDAY',
-        'MONDAY',
-        'TUESDAY',
-        'WEDNESDAY',
-        'THURSDAY',
-        'FRIDAY',
-        'SATURDAY',
-    ].map((day) => scheduleData[scheduleData[day]] || []);
     const now = new Date();
-
-    const parseTimeString = (timeStr) => {
-        const [time, modifier] = timeStr.trim().split(" ");
-        let [hours, minutes] = time.split(":").map(Number);
-
-        if (modifier.toUpperCase() === "PM" && hours !== 12) {
-            hours += 12;
-        } else if (modifier.toUpperCase() === "AM" && hours === 12) {
-            hours = 0;
-        }
-
-        const dateObj = new Date();
-        dateObj.setHours(hours);
-        dateObj.setMinutes(minutes);
-        dateObj.setSeconds(0);
-        dateObj.setMilliseconds(0);
-        return dateObj;
-    }
-
     const routeNames = Object.keys(routeData);
     const [selectedDay, setSelectedDay] = useState(now.getDay());
     const [selectedRoute, setSelectedRoute] = useState(routeNames[0]);
@@ -41,28 +14,7 @@ export default function Schedule() {
     const [schedule, setSchedule] = useState([]);
 
     useEffect(() => {
-        const scheduleByLoop = {};
-        Object.values(weeklySchedule[selectedDay]).forEach((busSchedule) => {
-            busSchedule.forEach(([time, loop]) => {
-                const timeObj = parseTimeString(time);
-                if (loop in scheduleByLoop) {
-                    scheduleByLoop[loop].push(timeObj);
-                } else {
-                    scheduleByLoop[loop] = [timeObj];
-                }
-            });
-        });
-        Object.values(scheduleByLoop).forEach((times) => {
-            times.sort((a, b) => {
-                const isA12AM = a.getHours() === 0 && a.getMinutes() === 0;
-                const isB12AM = b.getHours() === 0 && b.getMinutes() === 0;
-
-                if (isA12AM && !isB12AM) return 1;   // a goes after b
-                if (!isA12AM && isB12AM) return -1;  // a goes before b
-                return a - b;                        // otherwise, normal sort
-            });
-            });
-        setSchedule(scheduleByLoop);
+        setSchedule(parseSchedule(scheduleData, selectedDay));
     }, [selectedDay]);
 
     useEffect(() => {
